# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HelixWeave API",
    "description": "Reproducible omics workflows, from inputs to evidence.",
    "version": "0.4.1"
  },
  "paths": {
...
    "/api/v1/workflows/": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "List Workflows",
        "description": "List registered workflows with metadata and capabilities.",
        "operationId": "listWorkflows",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Workflow",
        "description": "Return one safe workflow product descriptor.",
        "operationId": "getWorkflow",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetailResponse"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetailResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetailResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/reference-profiles": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "List Compatible Reference Profiles",
        "description": "List only enabled exact revisions compatible with one workflow.",
        "operationId": "listCompatibleReferenceProfiles",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceProfileListResponse"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceProfileListResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceProfileListResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/schema": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Schema",
        "description": "Return the adapter-owned authoring contract for one workflow.",
        "operationId": "getWorkflowSchema",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaResponse"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/validate": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Validate Workflow",
        "description": "Validate submitted workflow inputs.",
        "operationId": "validateWorkflow",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body too large.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IssueResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "default": "error"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "technical_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technical Message"
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "IssueResponse",
        "description": "JSON-ready Issue shape from the PR84 contract."
      },
...
      "ReferenceProfileListResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/ReferenceProfileRevisionResponse"
            },
            "type": "array",
            "title": "Profiles"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ok",
          "workflow_id",
          "profiles"
        ],
        "title": "ReferenceProfileListResponse",
        "description": "Enabled compatible Reference Profiles for one workflow."
      },
...
      "SchemaResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "schema": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowSchemaResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "workflow_id",
          "schema"
        ],
        "title": "SchemaResponse",
        "description": "Envelope for GET /api/v1/workflows/{workflow_id}/schema."
      },
...
      "ValidatedInputSnapshotResponse": {
        "properties": {
          "snapshot_id": {
            "type": "string",
            "maxLength": 38,
            "pattern": "^vsnap_[0-9a-f]{32}$",
            "title": "Snapshot Id"
          },
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "adapter_version": {
            "type": "string",
            "title": "Adapter Version"
          },
          "payload_digest": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[0-9a-f]{64}$",
            "title": "Payload Digest"
          },
          "validated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Validated At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "project_id": {
            "type": "string",
            "maxLength": 36,
            "pattern": "^prj_[0-9a-f]{32}$",
            "title": "Project Id"
          },
          "binding_mode": {
            "type": "string",
            "enum": [
              "legacy_v1",
              "bound_v1"
            ],
            "title": "Binding Mode"
          },
          "provenance": {
            "type": "string",
            "enum": [
              "resolved",
              "unresolved"
            ],
            "title": "Provenance"
          },
          "sample_revision_ids": {
            "items": {
              "type": "string",
              "maxLength": 37,
              "pattern": "^smpr_[0-9a-f]{32}$"
            },
            "type": "array",
            "title": "Sample Revision Ids"
          },
          "binding_digest": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[0-9a-f]{64}$",
            "title": "Binding Digest"
          },
          "input_binding": {
            "$ref": "#/components/schemas/InputBindingResponse"
          },
          "reference_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReferenceProfileRevisionResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "snapshot_id",
          "workflow_id",
          "schema_version",
          "adapter_version",
          "payload_digest",
          "validated_at",
          "expires_at",
          "project_id",
          "binding_mode",
          "provenance",
          "sample_revision_ids",
          "binding_digest",
          "input_binding"
        ],
        "title": "ValidatedInputSnapshotResponse",
        "description": "Safe public projection of one server-owned validation snapshot."
      },
...
      "ValidationRequest": {
        "properties": {
          "config": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue-Input"
            },
            "type": "object",
            "title": "Config"
          },
          "samples": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4096,
                "minLength": 1
              },
              {
                "items": {
                  "additionalProperties": {
                    "type": "string",
                    "maxLength": 4096
                  },
                  "propertyNames": {
                    "maxLength": 128,
                    "minLength": 1
                  },
                  "type": "object",
                  "maxProperties": 64,
                  "minProperties": 1
                },
                "type": "array",
                "maxItems": 1000,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Samples"
          },
          "options": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue-Input"
            },
            "type": "object",
            "title": "Options"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 36,
                "pattern": "^prj_[0-9a-f]{32}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "sample_revision_ids": {
            "items": {
              "type": "string",
              "maxLength": 37,
              "pattern": "^smpr_[0-9a-f]{32}$"
            },
            "type": "array",
            "title": "Sample Revision Ids"
          },
          "input_selections": {
            "items": {
              "$ref": "#/components/schemas/InputFileRevisionSelectionRequest"
            },
            "type": "array",
            "maxItems": 256,
            "title": "Input Selections"
          },
          "reference_profile_revision_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 38,
                "pattern": "^refpr_[0-9a-f]{32}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Profile Revision Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "config"
        ],
        "title": "ValidationRequest",
        "description": "Request body for POST /api/v1/workflows/{workflow_id}/validate."
      },
      "ValidationResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id"
          },
          "value": {
            "type": "null",
            "title": "Value"
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidatedInputSnapshotResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "workflow_id",
          "snapshot"
        ],
        "title": "ValidationResponse",
        "description": "Envelope for POST /api/v1/workflows/{workflow_id}/validate."
      },
...
      "WorkflowCapabilityResponse": {
        "properties": {
          "supports": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supports"
          }
        },
        "type": "object",
        "title": "WorkflowCapabilityResponse",
        "description": "WorkflowCapabilities JSON shape."
      },
      "WorkflowDetailResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "workflow": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowListItem"
              },
              {
                "type": "null"
              }
            ]
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "workflow_id",
          "workflow"
        ],
        "title": "WorkflowDetailResponse",
        "description": "Envelope for GET /api/v1/workflows/{workflow_id}."
      },
...
      "WorkflowListItem": {
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/WorkflowMetadataResponse"
          },
          "schema_version": {
            "type": "string",
            "pattern": "^[1-9]\\d*\\.(?:0|[1-9]\\d*)\\.(?:0|[1-9]\\d*)$",
            "title": "Schema Version"
          },
          "capabilities": {
            "$ref": "#/components/schemas/WorkflowCapabilityResponse"
          },
          "upstream_identity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowUpstreamIdentityResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "availability": {
            "$ref": "#/components/schemas/WorkflowAvailabilityResponse"
          }
        },
        "type": "object",
        "required": [
          "metadata",
          "schema_version",
          "capabilities",
          "availability"
        ],
        "title": "WorkflowListItem",
        "description": "One entry in the workflow list response."
      },
      "WorkflowListResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "workflows": {
            "items": {
              "$ref": "#/components/schemas/WorkflowListItem"
            },
            "type": "array",
            "title": "Workflows"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "workflows"
        ],
        "title": "WorkflowListResponse",
        "description": "Envelope for GET /api/v1/workflows."
      },
...
    }
  }
}